packages/embcli-gemini/pyproject.toml
[project]
name = "embcli-gemini"
version = "0.0.1"
description = "gemini plugin for embcli"
readme = "README.md"
authors = [{ name = "Tomoko Uchida", email = "tomoko.uchida.1111@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "embcli-core",
    "google-genai>=1.14.0",
]

[tool.uv.sources]
embcli-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.entry-points."embcli"]
openai = "embcli_gemini.gemini"
